made some changes on useing entities
