+++ src/main/java/interface_adapters/friend/ViewFriendOutputAdapater.java
@@ -1,0 +1,14 @@
+package interface_adapters.friend;
+
+import entities.CommonUser;
+import usecases.friendinteractors.viewfriend.view_friend_output_boundary;
+
+import java.util.ArrayList;
+
+public class ViewFriendOutputAdapater implements view_friend_output_boundary {
+
+    @Override
+    public void pushFriendList(ArrayList<CommonUser> arraylist){
+        //comManager.send(arraylist)
+    }
+}
