+++ src/main/java/usecases/ProfileChanges/ChangeProfileInteractor.java
@@ -1,4 +1,4 @@
-package UseCases.ProfileChanges;
+package usecases.ProfileChanges;
 
 public class ChangeProfileInteractor {
 }
