+++ src/main/java/usecases/friendinteractors/addfriend/add_friend_DSGateway.java
@@ -1,6 +1,5 @@
 package usecases.friendinteractors.addfriend;
 
-import entities.CommonUser;
 
 public interface add_friend_DSGateway {
     void addFriendbyID(int requesterid, int friendid);
