+++ src/main/java/usecases/friendinteractors/deletefriend/delete_friend_input_boundary.java
@@ -1,7 +1,5 @@
 package usecases.friendinteractors.deletefriend;
 
-import usecases.friendinteractors.addfriend.add_friend_input_model;
-
 public interface delete_friend_input_boundary {
     public void DeleteFriend(delete_friend_input_model model);
 }
