+++ src/main/java/usecases/friendinteractors/viewfriend/view_friend_DSGateway.java
@@ -1,10 +1,8 @@
 package usecases.friendinteractors.viewfriend;
 
 import entities.CommonUser;
-
-import java.util.HashMap;
-import java.util.List;
+import java.util.ArrayList;
 
 public interface view_friend_DSGateway {
-    List<HashMap<String, Integer>> getFriendList();
+    ArrayList<CommonUser> getFriendList();
 }
