+++ src/main/java/usecases/friendinteractors/viewfriend/view_friend_input_boundary.java
@@ -1,7 +1,5 @@
 package usecases.friendinteractors.viewfriend;
 
-import usecases.friendinteractors.addfriend.add_friend_input_model;
-
 public interface view_friend_input_boundary {
     public void ViewFriend(view_friend_input_model model);
 }
