+++ src/main/java/usecases/friendinteractors/viewfriend/view_friend_input_model.java
